+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Change Log
 
+## [1.0.34] 2025-11-24
+
+### Added
+
+- Allow the -Debug switch on Invoke-HelmUpgrade to pass through to helm as --debug
+- Logging of pods that disappear from a ReplicaSet in Get-PodStatus
+
+### Changed
+
+- Verbose logging shows a line with the call stack for better debugging
+
+### Fixed
+
+- Path case in Deploy-Minimal.ps1 for Linux
+
 ## [1.0.33] 2025-07-14
 
 ### Changed